Copy full file timestamps
When the timestamp of the compressed file is copied from the source,
only the seconds where copied and the subsecond value was truncated to
zero.

This has the effect of making the newly created compressed file appear
to be older than the source file it was created from.

Use utimensat() call to set the full timestamps.  This makes the output
file have the identical timestamp as the input, which was the desired
behavior.